refactoring make slice from general context file
