+++ src/components/CountryList.jsx
@@ -2,7 +2,7 @@ import CountryItem from "./CountryItem";
 import styles from "./CountryList.module.css";
 import Spinner from "./Spinner";
 import Message from "./Message";
-import { useCities } from "../contexts/CitiesContext";
+import { useCities } from "../slices/citiesSlice/CitiesContext";
 
 function CountryList() {
   const { cities, isLoading } = useCities();
@@ -14,7 +14,7 @@ function CountryList() {
 
   function getVisitedCountries() {
     const countriesMap = cities.reduce((map, city) => {
-      const countryKey = `${city.country}-${city.emoji}`;
+      const countryKey = `${city.country}`;
       if (!map.has(countryKey)) {
         map.set(countryKey, { country: city.country, emoji: city.emoji });
       }
