--- src/contexts/CitiesContext.jsx
@@ -1,134 +1,0 @@
-import { createContext, useContext, useEffect, useReducer } from "react";
-import { BASE_URL } from "../EnvironmentVariables";
-
-const CityContext = createContext();
-
-const initialState = {
-  cities: [],
-  isLoading: false,
-  currentCity: {},
-  error: "",
-};
-
-function reduceCities(state, action) {
-  switch (action.type) {
-    case "loading":
-      return { ...state, isLoading: true };
-    case "cities/loaded":
-      return { ...state, isLoading: false, cities: action.payload };
-    case "city/loaded":
-      return { ...state, isLoading: false, currentCity: action.payload };
-    case "city/created":
-      return {
-        ...state,
-        isLoading: false,
-        cities: [...state.cities, action.payload],
-      };
-    case "city/deleted":
-      return {
-        ...state,
-        isLoading: false,
-        cities: state.cities.filter((city) => city.id !== action.payload),
-      };
-    case "rejected":
-      return { ...state, isLoading: false, error: action.payload };
-    default:
-      throw new Error("Something went wrong with cities data");
-  }
-}
-
-function CityContextProvider({ children }) {
-  const [{ cities, isLoading, currentCity }, dispatch] = useReducer(
-    reduceCities,
-    initialState
-  );
-
-  async function getCurrentCity(id) {
-    dispatch({ type: "loading" });
-    try {
-      const res = await fetch(`${BASE_URL}/cities/${id}`);
-      const data = await res.json();
-      dispatch({ type: "city/loaded", payload: data });
-    } catch (error) {
-      dispatch({ type: "rejected", payload: error.message });
-    }
-  }
-
-  useEffect(() => {
-    async function fetchCities() {
-      dispatch({ type: "loading" });
-      try {
-        const res = await fetch(`${BASE_URL}/cities`);
-        const data = await res.json();
-        dispatch({ type: "cities/loaded", payload: data });
-      } catch (error) {
-        dispatch({ type: "rejected", payload: error.message });
-      }
-    }
-    fetchCities();
-  }, []);
-
-  async function createCity(newCity) {
-    dispatch({ type: "loading" });
-    try {
-      const hasCity = cities.find(
-        (city) => city.cityName.toLowerCase() === newCity.cityName.toLowerCase()
-      );
-
-      if (hasCity) return;
-
-      const response = await fetch(`${BASE_URL}/cities`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(newCity),
-      });
-
-      if (!response.ok) {
-        throw new Error("Failed to create city.");
-      }
-      dispatch({ type: "city/created", payload: newCity });
-    } catch (error) {
-      dispatch({ type: "rejected", payload: error.message });
-    }
-  }
-
-  async function deleteCity(id) {
-    dispatch({ type: "loading" });
-    try {
-      const response = await fetch(`${BASE_URL}/cities/${id}`, {
-        method: "DELETE",
-      });
-
-      if (!response.ok) {
-        throw new Error("Failed to delete city.");
-      }
-      dispatch({ type: "city/deleted", payload: id });
-    } catch (error) {
-      dispatch({ type: "rejected", payload: error.message });
-    }
-  }
-
-  return (
-    <CityContext.Provider
-      value={{
-        cities,
-        isLoading,
-        currentCity,
-        getCurrentCity,
-        createCity,
-        deleteCity,
-      }}
-    >
-      {children}
-    </CityContext.Provider>
-  );
-}
-
-function useCities() {
-  const citiesContext = useContext(CityContext);
-  return citiesContext;
-}
-
-export { CityContextProvider, useCities };
